Renamed Ramiro analysis to composite analysis
Added deprecation warning and ensured backwards compatibility
