+++ pycqed/analysis/tomography.py
@@ -6,14 +6,16 @@ import scipy
 
 class TomoAnalysis_JointRO():
 
-    """Performs state tomography based on an overcomplete set of measurements
-     and calibration measurements. Uses qutip to calculate resulting basis states
-     from applied rotations.
-     
+    """
+    Performs state tomography based on an overcomplete set of measurements
+    and calibration measurements. Uses qutip to calculate resulting basis
+    states from applied rotations.
+
     Works for Joint RO (like REM experiment).
 
     Uses binary counting as general guideline in ordering states. Calculates
-     rotations by using the qutip library
+    rotations by using the qutip library
+
     BEFORE YOU USE THIS SET THE CORRECT ORDER BY CHANGING
         'rotation_matrixes'
         'measurement_basis' + 'measurement_basis_labels'
@@ -24,9 +26,10 @@ class TomoAnalysis_JointRO():
     # The set of single qubit rotation matrixes used in the tomography
     # measurement (will be assumed to be used on all qubits)
     rotation_matrixes = [qtp.identity(2), qtp.sigmax(),
-                         qtp.rotation(
-                             qtp.sigmax(), np.pi / 2), qtp.rotation(qtp.sigmay(), np.pi / 2),
-                         qtp.rotation(qtp.sigmax(), -np.pi / 2), qtp.rotation(qtp.sigmay(), -np.pi / 2)]
+                         qtp.rotation(qtp.sigmax(), np.pi / 2),
+                         qtp.rotation(qtp.sigmay(), np.pi / 2),
+                         qtp.rotation(qtp.sigmax(), -np.pi / 2),
+                         qtp.rotation(qtp.sigmay(), -np.pi / 2)]
     # The set of single qubit basis operators and labels
     measurement_basis = [
         qtp.identity(2), qtp.sigmaz(), qtp.sigmax(), qtp.sigmay()]
@@ -34,13 +37,17 @@ class TomoAnalysis_JointRO():
     # The operators used in the readout basis on each qubit
     readout_basis = [qtp.identity(2), qtp.sigmaz()]
 
-    def __init__(self, measurements_cal, measurements_tomo, n_qubits=2, n_quadratures=1):
+    def __init__(self, measurements_cal, measurements_tomo,
+                 n_qubits=2, n_quadratures=1):
         """
         keyword arguments:
         measurements_cal --- Should be an array of length 2 ** n_qubits
-        measurements_tomo --- Should be an array of length length(rotation_matrixes) ** n_qubits
-        n_qubits --- default(2) the amount of qubits present in the expirement
-        n_quadratures --- default(1(either I or Q)) The amount of complete measurement data sets. For example a combined IQ measurement has 2 measurement sets.
+        measurements_tomo --- Should be an array of length
+            length(rotation_matrixes) ** n_qubits
+        n_qubits --- default(2) the amount of qubits present in the experement
+        n_quadratures --- default(1(either I or Q)) The amount of complete
+            measurement data sets. For example a combined IQ measurement has
+            2 measurement sets.
 
         """
         self.measurements_cal = measurements_cal
@@ -59,11 +66,10 @@ class TomoAnalysis_JointRO():
         self.rotation_vector = self._calculate_matrix_set(
             self.rotation_matrixes, n_qubits)
 
-
-
     def execute_linear_tomo(self):
         """
-        Performs a linear tomography by simple inversion of the system of equations due to calibration points
+        Performs a linear tomography by simple inversion of the system of
+        equations due to calibration points
         """
 
         # calculate beta positions in coefficient matrix
@@ -75,20 +81,28 @@ class TomoAnalysis_JointRO():
         # re-add beta0
         basis_decomposition[0] = 1
         # now recreate the rho
-        rho = sum([basis_decomposition[i] * self.basis_vector[i] / (2 ** self.n_qubits)
+        rho = sum([basis_decomposition[i] * self.basis_vector[i] /
+                  (2 ** self.n_qubits)
                    for i in range(len(basis_decomposition))])
         return (basis_decomposition, rho)
 
-    def execute_max_likelihood(self, use_weights=True, show_time=True, ftol=0.01, xtol=0.001, full_output=0, max_iter=100):
+    def execute_max_likelihood(self, use_weights=True, show_time=True,
+                               ftol=0.01, xtol=0.001, full_output=0,
+                               max_iter=100):
         """
-        Performs a max likelihood optimization using fmin_powell in order to get the closest physically realisable state.
+        Performs a max likelihood optimization using fmin_powell in order to
+        get the closest physically realisable state.
 
-        This is done by constructing a lower triangular matrix T consisting of 4 ** n qubits params
+        This is done by constructing a lower triangular matrix T consisting of
+        4 ** n qubits params
+
         Keyword arguments:
-        use_weights : default(true) Weighs the quadrature data by the std in betas obtained
-        --- arguments for scipy fmin_powel method below, see the powel documentation
+        use_weights : default(true) Weighs the quadrature data by the std in
+                      betas obtained
+        --- arguments for scipy fmin_powel method below, see
+            the powel documentation
         """
-        # first we calculate the measurement matrixes
+        # first we calculate the measurement matrices
         tstart = time.time()
         measurement_vector = []
         n_rot = len(self.rotation_matrixes) ** self.n_qubits
@@ -96,28 +110,32 @@ class TomoAnalysis_JointRO():
         self.weights = np.ones(self.n_quadratures * n_rot)
         for quadrature in range(self.n_quadratures):
             betas = self._calibrate_betas(
-                self.measurements_cal[quadrature * self.n_states: (1 + quadrature) * self.n_states])
-            # determine the weights based on betas absulote difference and
+                self.measurements_cal[quadrature * self.n_states:
+                                      (1 + quadrature) * self.n_states])
+            # determine the weights based on betas absolote difference and
             # accuracy
             if (use_weights):
                 self.weights[
-                    quadrature * n_rot:(1+quadrature) * n_rot] = (max(betas) - min(betas)) / np.var(betas)
+                    quadrature * n_rot:(1+quadrature) * n_rot] = (
+                    max(betas) - min(betas)) / np.var(betas)
             for rotation_index, rotation in enumerate(self.rotation_vector):
                 measurement_vector.append(
-                    betas[0] * rotation.dag() * self.readout_vector[0] * rotation)
+                    betas[0] * rotation.dag()
+                    * self.readout_vector[0] * rotation)
                 for i in range(1, len(betas)):
                     measurement_vector[n_rot * quadrature + rotation_index] += betas[
                         i] * rotation.dag() * self.readout_vector[i] * rotation
         # save it in the object for use in optimization
         self.measurement_vector = measurement_vector
-        self.measurement_vector_numpy = [vec.full() for vec in measurement_vector]
+        self.measurement_vector_numpy = [
+            vec.full() for vec in measurement_vector]
         tlinear = time.time()
         # find out the starting rho by the linear tomo
         discard, rho0 = self.execute_linear_tomo()
         # now fetch the starting t_params from the cholesky decomp of rho
         tcholesky = time.time()
         T0 = np.linalg.cholesky((rho0.dag() * rho0).full() / 2)
-        t0 = np.zeros(4 ** self.n_qubits, dtype='complex' )
+        t0 = np.zeros(4 ** self.n_qubits, dtype='complex')
         di = np.diag_indices(2 ** self.n_qubits)
         tri = np.tril_indices(2 ** self.n_qubits, -1)
         t0[0:2 ** self.n_qubits] = T0[di]
@@ -126,14 +144,16 @@ class TomoAnalysis_JointRO():
         topt = time.time()
         # minimize the likelihood function using scipy
         t_optimal = scipy.optimize.fmin_powell(
-            self._max_likelihood_optimization_function, t0, maxiter=max_iter, full_output=full_output, ftol=ftol, xtol=xtol)
+            self._max_likelihood_optimization_function, t0, maxiter=max_iter,
+            full_output=full_output, ftol=ftol, xtol=xtol)
         if show_time is True:
             print(" Time to calc rotation matrixes %.2f " % (tlinear-tstart))
             print(" Time to do linear tomo %.2f " % (tcholesky-tlinear))
             print(" Time to build T %.2f " % (topt-tcholesky))
             print(" Time to optimize %.2f" % (time.time()-topt))
         return qtp.Qobj(self.build_rho_from_triangular_params(t_optimal),
-         dims=[[2 for i in range(self.n_qubits)], [2 for i in range(self.n_qubits)]])
+                        dims=[[2 for i in range(self.n_qubits)], [2 for i in
+                        range(self.n_qubits)]])
 
     def get_basis_labels(self, n_qubits):
         """
@@ -168,13 +188,17 @@ class TomoAnalysis_JointRO():
 
     def _max_likelihood_optimization_function(self, t_params):
         """
-        Optimization function that is evaluated many times in the maximum likelihood method.
-        Calculates the difference between expected measurement values and the actual measurement values based on a guessed rho
+        Optimization function that is evaluated many times in the maximum
+        likelihood method.
+
+        Calculates the difference between expected measurement values and the
+        actual measurement values based on a guessed rho
 
         Keyword arguments:
         t_params : cholesky decomp parameters used to construct the initial rho
         Requires:
-        self.weights :  weights per measurement vector used in calculating the loss
+        self.weights :  weights per measurement vector used in calculating the
+        loss
         """
         rho = self.build_rho_from_triangular_params(t_params)
         L = 0 + 0j
@@ -186,13 +210,16 @@ class TomoAnalysis_JointRO():
         return L
 
     def _calibrate_betas(self, measurements_cal):
-        """calculates betas from calibration points for the initial measurement operator
+        """
+        calculates betas from calibration points for the initial measurement
+        operator
 
         Betas are ordered by B0 -> II B1 -> IZ etc(binary counting)
         <0|Z|0> = 1, <1|Z|1> = -1
 
         Keyword arguments:
-        measurements_cal --- array(2 ** n_qubits) should be ordered correctly (00, 01, 10, 11) for 2 qubits
+        measurements_cal --- array(2 ** n_qubits) should be ordered
+            correctly (00, 01, 10, 11) for 2 qubits
         """
         cal_matrix = np.zeros((self.n_states, self.n_states))
         # get the coefficient matrix for the betas
@@ -206,18 +233,22 @@ class TomoAnalysis_JointRO():
 
     def _calculate_coefficient_matrix(self):
         """
-        Calculates the coefficient matrix used when inversing the linear system of equations needed to find rho
-        If there are multiple measurements present this will return a matrix of (n_quadratures * n_rotation_matrixes ** n_qubits) x n_basis_vectors
+        Calculates the coefficient matrix used when inversing the linear
+        system of equations needed to find rho
+        If there are multiple measurements present this will return a matrix
+        of (n_quadratures * n_rotation_matrixes ** n_qubits) x n_basis_vectors
         """
         coefficient_matrix = np.zeros(
-            (self.n_quadratures * len(self.rotation_matrixes) ** self.n_qubits, 4 ** self.n_qubits))
+            (self.n_quadratures * len(self.rotation_matrixes) ** self.n_qubits,
+             4 ** self.n_qubits))
         n_rotations = len(self.rotation_matrixes) ** self.n_qubits
         # Now fill in 2 ** self.n_qubits betas into the coefficient matrix on
         # each row
         for quadrature in range(self.n_quadratures):
             # calibrate betas for this quadrature
             self.betas = self._calibrate_betas(
-                self.measurements_cal[quadrature * self.n_states: (1 + quadrature) * self.n_states])
+                self.measurements_cal[quadrature * self.n_states:
+                (1 + quadrature) * self.n_states])
             for rotation_index in range(n_rotations):
                 for beta_index in range(2 ** self.n_qubits):
                     (place, sign) = self._get_basis_index_from_rotation(
@@ -227,8 +258,10 @@ class TomoAnalysis_JointRO():
         return coefficient_matrix
 
     def _get_basis_index_from_rotation(self, beta_index, rotation_index):
-        """Returns the position and sign of one of the betas in the coefficient matrix by checking
-        to which basis matrix the readout matrix is mapped after rotation
+        """
+        Returns the position and sign of one of the betas in the coefficient
+        matrix by checking to which basis matrix the readout matrix is mapped
+        after rotation
         This is used in _calculate_coefficient_matrix
         """
         m = self.rotation_vector[rotation_index].dag(
@@ -243,9 +276,12 @@ class TomoAnalysis_JointRO():
             'No basis vector found corresponding to the measurement rotation. Check that you have used Clifford Gates!')
 
     def _calculate_matrix_set(self, starting_set, n_qubits):
-        """recursive function that returns len(starting_set) ** n_qubits measurement_basis states tensored with eachother based on the amount of qubits
+        """recursive function that returns len(starting_set) ** n_qubits
+        measurement_basis states tensored with eachother based on the amount
+        of qubits
 
-        So for 2 qubits assuming your basis set is {I, X, Y, Z} you get II IX IY IZ XI XX XY XZ ...
+        So for 2 qubits assuming your basis set is {I, X, Y, Z}
+        you get II IX IY IZ XI XX XY XZ ...
         """
         if(n_qubits > 1):
             return [qtp.tensor(y, x) for x in self._calculate_matrix_set(starting_set, n_qubits - 1)
